refactor and fix coverage
