+++ src/index.ts
@@ -1,4 +1,4 @@
 /* eslint-disable import/no-unused-modules */
 
-export { AuthorityClient } from './lib/AuthorityClient.js';
-export type { AuthorizationHeader } from './lib/AuthorizationHeader.js';
+export { AuthorityClient } from './lib/client/AuthorityClient.js';
+export type { AuthorizationHeader } from './lib/client/AuthorizationHeader.js';
