+++ src/lib/AuthorityError.ts
@@ -1,0 +1,5 @@
+export abstract class AuthorityError extends Error {
+  public override get name(): string {
+    return this.constructor.name;
+  }
+}
