+++ src/lib/client/ClientError.ts
@@ -1,0 +1,6 @@
+import { AuthorityError } from '../AuthorityError.js';
+
+/**
+ * Represents a 4XX response.
+ */
+export class ClientError extends AuthorityError {}
